requirements.txt
# Environment and config
python-dotenv

# LangChain core
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters

# Vector store
faiss-cpu

# Document loaders & processing
pypdf
python-docx
tiktoken
typing-extensions
